fix auto update

+++ index.ts
@@ -30,6 +30,7 @@ function download(url: string, path: string, cb: any = undefined) {
 	const file = fs.createWriteStream(path);
 	try {
 		const request = http.get(url, (response) => {
+			if (response.statusCode !== 200) return false;
 			response.pipe(file);
 		});
 
@@ -41,12 +42,19 @@ function download(url: string, path: string, cb: any = undefined) {
 			fs.unlink(path, () => cb(err.message));
 		});
 	} catch {
-		throw "";
+		return false;
 	};
+
+	return true;
 };
 
 async function update() {
-	await download("http://CIF.kro.kr/398znmfl-rf-zrekip029z-qwerwe/zmofip=43-8900ua34j3-09-124825425234-z9i90j/CIF.zip", "../plugins/CIF.zip");
+	if(await download("http://CIF.kro.kr/398znmfl-rf-zrekip029z-qwerwe/zmofip=43-8900ua34j3-09-124825425234-z9i90j/CIF.zip", "../plugins/CIF.zip") !== true) {
+		import("./modules/util/configManager");
+		import("./main");
+		return;
+	};
+
 	await exec('cmd /c rmdir /q /s "../plugins/cif"', ((err, stdout, stderr) => {
 		if (err) throw err;
 	}));
